--- packages/shared/src/components/index.ts
@@ -1,6 +1,0 @@
-export * from "./CaptionPopover";
-export * from "./CardError";
-export * from "./CardLoading";
-export * from "./visualizations";
-
-
