--- packages/shared/src/components/visualizations/index.ts
@@ -1,15 +1,0 @@
-export * from "./ChartVisualizer";
-export * from "./CustomOrgUnitModal";
-export * from "./CustomPeriodModal";
-export * from "./DataVisComponent";
-export * from "./MapVisComponent";
-export * from "./Banner";
-export * from "./ActionMenu";
-export * from "./YearOverYearDataVisComponent";
-export * from "./VisualizationTitle";
-export * from "./PDFVisualizer";
-export * from "./ChartSelector";
-export * from "./MapTableComponent";
- 
-
-
