+++ packages/ui/src/visualizations/YearOverYearVisualizer.tsx
@@ -1,5 +1,5 @@
 import { AnalyticsData, VisualizationConfig } from "@packages/shared/schemas";
-import { RefObject, useRef } from "react";
+import React, { RefObject, useRef } from "react";
 import HighchartsReact from "highcharts-react-official";
 
 export interface ChartVisualizerProps {
