+++ packages/ui/src/visualizations/index.ts
@@ -3,3 +3,5 @@ export * from "./TableVisualizer";
 export * from "./SingleValueVisualizer";
 export * from "./RichTextView";
 export * from "./MapVisualizer";
+export * from "./VisualizationTitle";
+export * from "./ChartSelector";
